Move UniqueIdSupplier(s) from styx-api module into styx-server module.
