+++ components/server/src/main/java/com/hotels/styx/server/netty/codec/NettyToStyxRequestDecoder.java
@@ -18,7 +18,7 @@ package com.hotels.styx.server.netty.codec;
 import com.google.common.annotations.VisibleForTesting;
 import com.hotels.styx.api.HttpCookie;
 import com.hotels.styx.api.HttpMessageBody;
-import com.hotels.styx.api.UniqueIdSupplier;
+import com.hotels.styx.server.UniqueIdSupplier;
 import com.hotels.styx.api.Url;
 import com.hotels.styx.server.BadRequestException;
 import io.netty.buffer.ByteBuf;
@@ -53,7 +53,7 @@ import static com.google.common.collect.Iterables.size;
 import static com.hotels.styx.api.HttpHeaderNames.COOKIE;
 import static com.hotels.styx.api.HttpHeaderNames.EXPECT;
 import static com.hotels.styx.api.HttpHeaderNames.HOST;
-import static com.hotels.styx.api.UniqueIdSuppliers.UUID_VERSION_ONE_SUPPLIER;
+import static com.hotels.styx.server.UniqueIdSuppliers.UUID_VERSION_ONE_SUPPLIER;
 import static com.hotels.styx.api.Url.Builder.url;
 import static com.hotels.styx.api.common.Strings.quote;
 import static com.hotels.styx.api.support.CookiesSupport.isCookieHeader;
